backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
reportlab
httpx==0.27.0
pydantic-settings==2.4.0
python-dotenv==1.0.1
pdfminer.six==20220524
python-docx==0.8.11


